+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * [CodeReloader] The `lib/` directory is now code reloaded by default along with `web/` in development
   * [Channel] Add `subscribe/2` and `unsubscribe/2` to handle external topic subscriptions for a socket
   * [View] Generate private `render_template/2` clauses for views to allow overriding `render/2` clauses before rendering templates
+  * [Channel] Add `:phoenix_channel_join` instrumentation hook
 
 * Deprecations
   * [Endpoint] Generated `subscribe/3` and `unsubscribe/2` clauses have been deprecated in favor of `subscribe/2` and `unsubscribe/1` which uses the caller's pid